+++ src/components/alert.js
@@ -1,5 +1,4 @@
-import {createElement} from "../utils";
-
+import AbstractComponent from './abstract-class.js';
 const createAlertTemplate = () => {
   return (
     `<section class="films">
@@ -11,20 +10,8 @@ const createAlertTemplate = () => {
 };
 
 
-export default class Alert {
-  constructor() {
-    this._element = null;
-  }
+export default class Alert extends AbstractComponent {
   getTemplate() {
     return createAlertTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
