+++ src/components/show-more-button.js
@@ -1,0 +1,16 @@
+import AbstractComponent from "./abstract-class";
+
+const createShowMoreButtonTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export default class ShowMoreButton extends AbstractComponent {
+  getTemplate() {
+    return createShowMoreButtonTemplate();
+  }
+  setClickHandler(handler) {
+    this.getElement().addEventListener(`click`, handler);
+  }
+}
