+++ src/components/sort-form.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-class";
 
 const createSortFormTemplate = () => {
   return (
@@ -10,20 +10,8 @@ const createSortFormTemplate = () => {
   );
 };
 
-export default class SortForm {
-  constructor() {
-    this._element = null;
-  }
+export default class SortForm extends AbstractComponent {
   getTemplate() {
     return createSortFormTemplate();
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
   }
 }
